requirements.txt
pygame==2.5.2
pymongo==4.6.1
mysql-connector-python==8.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
Flask==3.0.0
flask-cors==4.0.0
cryptography==41.0.7
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
bcrypt==4.1.2
